store and load config from redis
closes #83
